+++ src/webviews/index.ts
@@ -1,5 +1,4 @@
 export { PasswordGeneratorProvider } from './passwordGeneratorProvider';
 export { Icons } from './icons';
-export { NotesPanel } from './notesPanel';
 export { NotesTableProvider } from './notesTableProvider';
 export { NoteEditorProvider } from './noteEditorProvider';